Various cleanup: warning fixes, improvements to function names, better debugging commands
Types
Changed int to size_t and size_t to int to remove warnings.
Explicitly cast float to int to remove warnings.
Explicitly case size() to int to remove warnings.

Clearer function names
Changed UCI::move and UCI::square to UCI::to_string(Move) and UCI::to_string(Square).
Changed Position::set to Position::init.
Removed Utils::lcm
Removed using namespace std.

Enhancement
Added more info to board ASCII printout (move number, turn and declare check)
Added debugging commands "eval" and "testgame".
Improvements to debugging commands.
Debug command 'play' now accepts a sequence of moves.
Debug command 'showgame' now emits a string that recreates the board state, including the last 8 moves.
Checkmate/Stalemate/Draw/Check now properly declared when moving with 'go' command.
Added Position::has_no_moves.

Bugfixes
Crash when 'go' command issued at end of game
Missing underscore WIN32

Project settings
Moved the shared defines to the left.

+++ src/Timing.cpp
@@ -22,8 +22,8 @@
 
 
 int Time::timediff_centis(Time start, Time end) {
-    return std::chrono::duration_cast<std::chrono::milliseconds>
-        (end.m_time - start.m_time).count() / 10;
+    return int(std::chrono::duration_cast<std::chrono::milliseconds>
+        (end.m_time - start.m_time).count() / 10);
 }
 
 double Time::timediff_seconds(Time start, Time end) {
